requirements.txt
asgiref
Django
psycopg2-binary
sqlparse
django-crispy-forms
crispy-bootstrap5
django-allauth